pyproject.toml
[project]
name = "tszpaint"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "abacusutils[all]>=2.1.2",
    "asdf>=5.1.0",
    "astropy>=7.1.1",
    "cosmo>=0.7.0",
    "h5py>=3.15.1",
    "healpy>=1.19.0",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "jax>=0.8.0",
    "loguru>=0.5.3",
    "matplotlib>=3.10.7",
    "numba>=0.63.1",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "psutil>=7.2.2",
    "quadax>=0.2.11",
    "ruff>=0.15.1",
    "scikit-learn>=1.8.0",
    "scipy-stubs>=1.17.0.2",
    "unxt>=1.7.7",
    "wcosmo>=0.5.4",
    "wrapt>=2.1.1",
]

[dependency-groups]
dev = [
    "ipython>=9.7.0",
]
